% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinoperatorqubit.R
\name{spin_operator_qubit}
\alias{spin_operator_qubit}
\title{illustration des opérations de porte quantique avec des qubits de spin}
\usage{
spin_operator_qubit(spin_values, quantum_gate)
}
\arguments{
\item{spin_values}{La liste de valeurs de spin}

\item{quantum_gate}{La valeur de la porte quantique}
}
\value{
La matrice qubits sous forme de vecteur
}
\description{
illustration des opérations de porte quantique avec des qubits de spin
}
\examples{
spin_operator_qubit(spin_values=c(0.5, -0.5, 0.8, -0.3), quantum_gate = "Hadamard")

}
